♻️ refactor: 기존의 DB 연결 코드를 하나의 모듈로 통합 #18
- DB_setting 모듈에서 하나로 통합
- return 타입은 conn, cur을 가져와서 사용
